backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
beautifulsoup4
playwright
python-dotenv
